refactor: move App.tsx file in app folder. move styles and d.ts files in app folder
