+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import './index.css';
-import App from './App.tsx';
+import './app/styles/index.css';
+import App from './app/App.tsx';
 
 const rootElement = document.getElementById('root');
 if (rootElement) {
